requirements.txt
Django
psycopg2>=2.8
pillow
python-dotenv
djangorestframework
django-rest-swagger
setuptools
drf-spectacular
django-cors-headers